OpenAeroVTOL Beta 55.2
Beta 54	
 - Revert to B46-era IMU. 
 - Read eeProm back into config structure.
 - Made text-based values increment more slowly. 
 - LED flashes to show successful acc calibrations
 - I-rates now default to 2. 
 - Checked that all defaults are the same as factory settings.
 - Fixed pbuffer overrun which caused text to be centered off-screen.
 - Removed advanced features to simplify debugging.
 - Wound back LCD_Display_Text() indexing to B46, ServoOut[] no longer volatile

Beta 55
 - Bugfix release for IMU that could no longer handle tumbling.
 - Had to reinstate the use of TMR0 to count the over-runs of TMR1 when measuring the loop interval.

Beta 55.1
 - Still buggy for weird AL on existing the menu. Trying new loop interval handling.

Beta 55.2
 - Bug fixed. Minor tweaks. Limit the instantaneous change in angle of the IMU to +/-15 degrees.
 - Removed read-back of eeProm into config structure.
 - Updated the default eeprom settings. Scrimped a few bytes to squeeze it into the KK2.0.

+++ OpenAeroVTOL/OpenAeroVTOL_2_0/inc/glcd_menu.h
@@ -6,7 +6,7 @@
 //* Externals
 //***********************************************************
 
-extern void LCD_Display_Text (uint16_t menuitem, const unsigned char* font,uint16_t x, uint16_t y);
+extern void LCD_Display_Text (uint8_t menuitem, const unsigned char* font,uint16_t x, uint16_t y);
 extern void gLCDprint_Menu_P(const char *s, const unsigned char* font,uint16_t x, uint16_t y);
